+++ lib/ui/settings/SettingsHomeContentPanel.dart
@@ -52,7 +52,7 @@ class SettingsHomeContentPanel extends StatefulWidget {
 
   static void present(BuildContext context, {SettingsContent? content}) {
     if (ModalRoute.of(context)?.settings.name != routeName) {
-      MediaQueryData mediaQuery = MediaQueryData.fromWindow(WidgetsBinding.instance.window);
+      MediaQueryData mediaQuery = MediaQueryData.fromView(View.of(context));
       double height = mediaQuery.size.height - mediaQuery.viewPadding.top - mediaQuery.viewInsets.top - 16;
       showModalBottomSheet(
         context: context,