+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add survey creation tool [#3295](https://github.com/rokwire/illinois-app/issues/3295).
 ### Fixed
-- Upgrade dependencies for Flutter v3.10 [#3294](https://github.com/rokwire/illinois-app/issues/3294)
+- Upgrade dependencies for Flutter v3.10 [#3294](https://github.com/rokwire/illinois-app/issues/3294).
+### Changed
+- Cleaned up Explore interface [#3301](https://github.com/rokwire/illinois-app/issues/3301)
 
 ## [5.0.5] - 2023-05-23
 - Created and acknowledged at different places async versions of JSON encode/decode and collection equality checks [#3292](https://github.com/rokwire/illinois-app/issues/3292).