+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [5.0.6] - 2023-05-29
 ### Added
 - Add survey creation tool [#3295](https://github.com/rokwire/illinois-app/issues/3295).
 ### Fixed